Fix type annotation for max_delay parameter in measure_time function and add .gitignore for __pycache__ directory

+++ python_async_function/2-measure_runtime.py
@@ -7,7 +7,7 @@ import time
 wait_n = __import__('1-concurrent_coroutines').wait_n
 
 
-def measure_time(n: int, max_delay: float) -> float:
+def measure_time(n: int, max_delay: int) -> float:
     """Measure average execution time for wait_n coroutines."""
     start = time.time()
     asyncio.run(wait_n(n, max_delay))
